refactor to condense code and update tests.
