--- packages/editor/src/store/actions/index.js
@@ -1,2 +1,0 @@
-export * from './post-actions';
-export * from './post-generators';
