+++ packages/editor/src/store/index.js
@@ -9,16 +9,14 @@ import { registerStore } from '@wordpress/data';
 import reducer from './reducer';
 import applyMiddlewares from './middlewares';
 import * as selectors from './selectors';
-import * as generalActions from './actions.js';
-import * as actions from './actions/index.js';
+import * as actions from './actions';
 import controls from './controls';
 import { MODULE_KEY } from './constants';
 
 const store = registerStore( MODULE_KEY, {
 	reducer,
-	controls,
 	selectors,
-	actions: { ...actions, ...generalActions },
+	actions,
 	controls,
 	persist: [ 'preferences' ],
 } );
